Add a bunch more context managers
- compilation
- testSuite
- test
- progress
- serviceMessagesDisabled
- serviceMessagesEnabled

Context managers are implemented using `contextlib.contextmanager`, as
using `contextlib.contextmanager` is much easier than creating a context
manager class.

I do it in a separate file that gets imported on-demand at run-time, so
that we remain compatible with Python 2.4, which doesn't have the
`yield` statement (or context managers).

This way, folks can use the context managers in Python >= 2.5 and in
Python 2.4, they can't use the context managers, but the rest of the
library works.