Add escape_value function
In some cases, it might be useful to call escape_value as a simple
function without instantiating a `TeamcityServiceMessages` object.

e.g.:

    In [1]: from teamcity.messages import escape_value

    In [2]: escape_value("That's some text in [square brackets] and |vertical bars|\n")
    Out[2]: "That|'s some text in [square brackets|] and ||vertical bars|||n"

This is useful to me, because I have some bash code that emits TeamCity
messages and I would like to be able to reuse this Python code to do the
escaping. With this, I can change the code from:

    teamcity_escape() {
      echo "$@" | python -c 'import sys; from teamcity.messages import TeamcityServiceMessages; tc = TeamcityServiceMessages(); print(tc.escapeValue(sys.stdin.read()))'
    }

to:

    teamcity_escape() {
      echo "$@" | python -c 'import sys; from teamcity.messages import escape_value; print(tc.escape_value(sys.stdin.read()))'
    }